changed method of hiding fields from overriding pages to using javascript
Removed unnecessary files and added a hook file with logic to build
javascript code from prototype framework to hide/show fields based on
the role permissions. The code is not the greatest and needs to be
improved as I am still a novice Ruby programmer. Hope to making
improvements as time allows.